Cleaned up some code
