+++ medication-reminder/CustomDate.swift
@@ -40,23 +40,3 @@ class CustomDate: Hashable {
     }
     
 }
-
-extension Date {
-    static let iso8601Formatter: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.calendar = Calendar(identifier: .iso8601)
-        formatter.locale = Locale(identifier: "en_US_POSIX")
-        formatter.timeZone = TimeZone(identifier: "UTC")
-        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSXXXXX"
-        return formatter
-    }()
-    var iso8601: String {
-        return Date.iso8601Formatter.string(from: self)
-    }
-}
-
-extension String {
-    var dateFromISO8601: Date? {
-        return Date.iso8601Formatter.date(from: self)
-    }
-}
